refactor: DayBox 컴포넌트 분리
