+++ src/pages/admin-MainPage/SchedulePage/FixedMonthlyConents.tsx
@@ -3,16 +3,8 @@ import { getMonthly } from 'apis/getSchedule';
 import { useAtom } from 'jotai';
 import { dateAtom, monthAtom } from 'pages/admin-MainPage/SchedulePage/AdminScheduleSection';
 import React from 'react';
-import {
-  Badge,
-  BadgeCont,
-  DateCont,
-  InnerDayBox,
-  MonthBox,
-  OutterDayBox,
-  WeekGrid,
-} from '../../../components/@commons-feature/calendar/CalendarStyle';
-import Text from 'components/@commons/Text';
+import { MonthBox, WeekGrid } from 'components/@commons-feature/calendar/CalendarStyle';
+import FixedDailyBox from './FixedDailyBox';
 
 interface Props {
   selectedId: number;
@@ -22,7 +14,7 @@ interface DailyData {
   workTime: string[] | null;
 }
 
-const MonthlyInner = ({ selectedId }: Props): JSX.Element => {
+const FixedMonthlyConents = ({ selectedId }: Props): JSX.Element => {
   const [selectedDate, setSelectedDate] = useAtom(dateAtom);
   const [nowMonth] = useAtom(monthAtom);
 
@@ -47,7 +39,7 @@ const MonthlyInner = ({ selectedId }: Props): JSX.Element => {
       {scheduleData?.table.map((weekArray: DailyData[], i) => (
         <WeekGrid key={`${i}주`}>
           {weekArray.map((e: DailyData) => (
-            <DayBox
+            <FixedDailyBox
               key={e.date}
               dateString={e.date}
               timeList={e.workTime}
@@ -61,35 +53,4 @@ const MonthlyInner = ({ selectedId }: Props): JSX.Element => {
   );
 };
 
-export default MonthlyInner;
-
-interface DayBoxProps {
-  dateString: string;
-  timeList: string[] | null;
-  onClick: React.MouseEventHandler<HTMLDivElement>;
-  isSelected: boolean;
-}
-
-const DayBox = ({ dateString, timeList, onClick, isSelected }: DayBoxProps): JSX.Element => {
-  const date = dateString.split('-').map((e) => Number.parseInt(e))[2];
-
-  return (
-    <OutterDayBox onClick={onClick} $disabled={timeList === null}>
-      <InnerDayBox $isSelected={isSelected}>
-        <DateCont $isToday={date === new Date().getDate()}>
-          <Text size="xs" weight="regular">
-            {date}
-          </Text>
-        </DateCont>
-        <BadgeCont>
-          {!!timeList &&
-            timeList.map((t) => (
-              <Badge key={t} $time={t}>
-                <Text size="sm">{t}</Text>
-              </Badge>
-            ))}
-        </BadgeCont>
-      </InnerDayBox>
-    </OutterDayBox>
-  );
-};
+export default FixedMonthlyConents;
